+++ lib/ga4gh/example/passport_broker.go
@@ -15,11 +15,13 @@
 package main
 
 import (
+	"context"
 	"fmt"
 	"time"
 
 	"github.com/pborman/uuid" /* copybara-comment */
 	"github.com/GoogleCloudPlatform/healthcare-federated-access-services/lib/ga4gh" /* copybara-comment: ga4gh */
+	"github.com/GoogleCloudPlatform/healthcare-federated-access-services/lib/kms/localsign" /* copybara-comment: localsign */
 	"github.com/GoogleCloudPlatform/healthcare-federated-access-services/lib/testkeys" /* copybara-comment: testkeys */
 )
 
@@ -45,9 +47,11 @@ func (b *PassportBroker) FetchAccess(t Token) (ga4gh.AccessJWT, error) {
 		},
 	}
 
-	p, err := ga4gh.NewAccessFromData(d, ga4gh.RS256, b.Key.Private, "kid")
+	ctx := context.Background()
+	signer := localsign.New(&b.Key)
+	p, err := ga4gh.NewAccessFromData(ctx, d, signer)
 	if err != nil {
-		return "", fmt.Errorf("NewAccessFromData(%v,%v,%v) failed:\n%v", d, ga4gh.RS256, b.Key.Private, err)
+		return "", fmt.Errorf("NewAccessFromData() failed:\n%v", err)
 	}
 	return p.JWT(), nil
 }
@@ -65,11 +69,6 @@ func (b *PassportBroker) FetchVisas(t Token) ([]ga4gh.VisaJWT, error) {
 	v, err := ga4gh.NewVisaFromJWT(j)
 	if err != nil {
 		return nil, fmt.Errorf("NewVisaFromJWT(%v) failed:\n%v", j, err)
-	}
-
-	// Optional
-	if err := v.Verify(b.I.Key.Public); err != nil {
-		return nil, fmt.Errorf("Visa(%v).Verify(%v) failed:\n%v", v, b.I.Key.Public, err)
 	}
 
 	return []ga4gh.VisaJWT{v.JWT()}, nil