+++ lib/ga4gh/identity.go
@@ -19,8 +19,8 @@ import (
 	"fmt"
 	"net/url"
 	"strings"
-	"time"
 
+	"github.com/GoogleCloudPlatform/healthcare-federated-access-services/lib/errutil" /* copybara-comment: errutil */
 	"github.com/GoogleCloudPlatform/healthcare-federated-access-services/lib/globalflags" /* copybara-comment: globalflags */
 	"github.com/GoogleCloudPlatform/healthcare-federated-access-services/lib/httputils" /* copybara-comment: httputils */
 	"github.com/GoogleCloudPlatform/healthcare-federated-access-services/lib/strutil" /* copybara-comment: strutil */
@@ -119,33 +119,6 @@ type Identity struct {
 	Extra            map[string]interface{} `json:"ext,omitempty"`
 }
 
-// Valid implements dgrijalva/jwt-go Claims interface. This will be called when using
-// dgrijalva/jwt-go parse. This validates exp, iat, nbf in token.
-func (t *Identity) Valid() error {
-	return t.Validate("")
-}
-
-// Validate returns an error if the Identity does not pass basic checks.
-func (t *Identity) Validate(clientID string) error {
-	now := time.Now().Unix()
-
-	if now > t.Expiry {
-		return fmt.Errorf("token is expired")
-	}
-
-	if now < t.IssuedAt {
-		return fmt.Errorf("token used before issued")
-	}
-
-	if now < t.NotBefore {
-		return fmt.Errorf("token is not valid yet")
-	}
-
-	// TODO: check non-empty clientID against t.Audiences
-
-	return nil
-}
-
 // CheckIdentityAllVisasLinked checks if the Visas inside the identity are linked.
 // Verifies all Visas of type LinkedIdentities.
 // If JWTVerifier is not nil, will call f to verify LinkedIdentities Visas.
@@ -159,7 +132,7 @@ func CheckIdentityAllVisasLinked(ctx context.Context, i *Identity, f JWTVerifier
 		}
 
 		if f != nil && v.Data().Assertion.Type == LinkedIdentities {
-			if err := f(ctx, j); err != nil {
+			if err := f(ctx, j, v.Data().Issuer, v.JKU()); err != nil {
 				return fmt.Errorf("the verification of some LinkedIdentities visa failed: %v", err)
 			}
 		}
@@ -223,8 +196,12 @@ func VisasToOldClaims(ctx context.Context, visas []VisaJWT, f JWTVerifier) (map[
 		}
 
 		if f != nil {
-			if err := f(ctx, string(j)); err != nil {
-				rejected = append(rejected, NewRejectedVisa(d, v.Format(), "verify_failed", "", err.Error()))
+			if err := f(ctx, string(j), v.Data().Issuer, v.JKU()); err != nil {
+				reason := errutil.ErrorReason(err)
+				if len(reason) == 0 {
+					reason = "verify_failed"
+				}
+				rejected = append(rejected, NewRejectedVisa(d, v.Format(), reason, "", err.Error()))
 				continue
 			}
 		}